pyproject.toml
[project]
name = "expert-agent"
version = "0.1.0"
description = "Ultra-specialist AI agents as a service — NotebookLM-style, powered by Gemini 3.1 Pro long-context + caching."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Felipe Broering", email = "hi@felipe.run" }]
keywords = ["ai", "agents", "gemini", "llm", "notebooklm", "cloud-run", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "structlog>=24.4",
    "google-genai>=0.3.0",
    "google-cloud-firestore>=2.19",
    "google-cloud-storage>=2.18",
    "google-cloud-secret-manager>=2.21",
    "mempalace>=0.1.0",
    "chromadb-client>=0.5",
    "httpx>=0.27",
    "tenacity>=9.0",
    "slowapi>=0.1.9",
    "bcrypt>=4.2",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1",
    "prometheus-client>=0.21",
    "typer>=0.15",
    "rich>=13.9",
]

[project.optional-dependencies]
vertex = [
    "google-cloud-aiplatform>=1.72",
]
otel = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-gcp-trace>=1.7",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "respx>=0.21",
    "mock-firestore>=0.11",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-PyYAML",
]

[project.scripts]
agent-cli = "agent_cli.main:app"
expert-agent-backend = "app.main:run"

[project.urls]
Homepage = "https://github.com/feliperbroering/expert-agent"
Repository = "https://github.com/feliperbroering/expert-agent"
Issues = "https://github.com/feliperbroering/expert-agent/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Deploy target is Linux/amd64 (Cloud Run). Some transitive deps (e.g. onnxruntime
# via chromadb) don't have wheels for every platform (notably macOS x86_64). Telling uv
# the required environments ensures the lockfile always resolves for our prod target.
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/app", "cli/agent_cli"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W",  # pycodestyle + pyflakes
    "I",            # isort
    "B",            # bugbear
    "UP",           # pyupgrade
    "N",            # pep8-naming
    "SIM",          # simplify
    "RUF",
]
ignore = ["E501"]  # handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "N802"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = "backend:cli"
exclude = ["tests/", "build/", "dist/"]

[[tool.mypy.overrides]]
module = ["mempalace.*", "chromadb.*", "mock_firestore.*", "slowapi.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --cov-report=term-missing --import-mode=importlib"
testpaths = ["backend/tests", "cli/tests"]
pythonpath = ["backend", "cli"]

[tool.coverage.run]
source = ["backend/app", "cli/agent_cli"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]
